+++ NEWS.md
@@ -2,6 +2,8 @@
 
 * Fixed unstable behaviour of regex in C++ on both the psoho and natPsoho algorithms.
 
+* Added a whitelist_tr argument to the DMMHC algorithm to allow forcing arcs in the transition network.
+
 # dbnR 0.7.8
 
 * Bringing the package overall status up to scratch for publication on a journal.